Pythagoras tree: Refactored out classes into logical files
