+++ orangecontrib/prototypes/widgets/pythagorastreeviewer.py
@@ -14,22 +14,15 @@ Square : namedtuple (center, length, angle)
 Point : namedtuple (x, y)
     Self exaplanatory.
 
-Classes
--------
-PythagorasTreeViewer
-SquareGraphicsItem
-
 """
-from collections import namedtuple, defaultdict, deque, OrderedDict
-from functools import lru_cache
+from collections import namedtuple, defaultdict, deque
 from math import pi, sqrt, cos, sin, degrees
 
-import numpy as np
-from Orange.preprocess.transformation import Indicator
+from orangecontrib.prototypes.utils.tree.treeadapter import TreeAdapter
+
 from PyQt4 import QtCore, QtGui
 from PyQt4.QtCore import Qt
 
-# Please note that all angles are in radians
 # z index range, increase if needed
 Z_STEP = 5000000
 
@@ -391,30 +384,27 @@ class InteractiveSquareGraphicsItem(SquareGraphicsItem):
 
         self.setToolTip(kwargs.get('tooltip', 'Tooltip'))
 
-        InteractiveSquareGraphicsItem.timer.setSingleShot(True)
-
-    def _scene_selection_manager(self):
-        pass
+        self.timer.setSingleShot(True)
 
     def hoverEnterEvent(self, ev):
-        InteractiveSquareGraphicsItem.timer.stop()
+        self.timer.stop()
 
         def fnc(graphics_item):
             graphics_item.setZValue(Z_STEP)
             if self.any_selected:
                 if graphics_item.isSelected():
-                    opacity = InteractiveSquareGraphicsItem.MAX_OPACITY
+                    opacity = self.MAX_OPACITY
                 else:
-                    opacity = InteractiveSquareGraphicsItem.SELECTION_OPACITY
+                    opacity = self.SELECTION_OPACITY
             else:
-                opacity = InteractiveSquareGraphicsItem.MAX_OPACITY
+                opacity = self.MAX_OPACITY
             graphics_item.setOpacity(opacity)
 
         def other_fnc(graphics_item):
             if graphics_item.isSelected():
-                opacity = InteractiveSquareGraphicsItem.MAX_OPACITY
+                opacity = self.MAX_OPACITY
             else:
-                opacity = InteractiveSquareGraphicsItem.HOVER_OPACITY
+                opacity = self.HOVER_OPACITY
             graphics_item.setOpacity(opacity)
             graphics_item.setZValue(self.initial_zvalue)
 
@@ -430,17 +420,17 @@ class InteractiveSquareGraphicsItem(SquareGraphicsItem):
         def other_fnc(graphics_item):
             if self.any_selected:
                 if graphics_item.isSelected():
-                    opacity = InteractiveSquareGraphicsItem.MAX_OPACITY
+                    opacity = self.MAX_OPACITY
                 else:
-                    opacity = InteractiveSquareGraphicsItem.SELECTION_OPACITY
+                    opacity = self.SELECTION_OPACITY
             else:
-                opacity = InteractiveSquareGraphicsItem.MAX_OPACITY
+                opacity = self.MAX_OPACITY
             graphics_item.setOpacity(opacity)
 
-        InteractiveSquareGraphicsItem.timer.timeout.connect(
+        self.timer.timeout.connect(
             lambda: self._propagate_z_values(self, fnc, other_fnc))
 
-        InteractiveSquareGraphicsItem.timer.start(250)
+        self.timer.start(250)
 
     def _propagate_z_values(self, graphics_item, fnc, other_fnc):
         self._propagate_to_children(graphics_item, fnc)
@@ -475,16 +465,14 @@ class InteractiveSquareGraphicsItem(SquareGraphicsItem):
                 dropshadow.setColor(QtGui.QColor(Qt.black))
                 dropshadow.setOffset(0, 0)
                 self.setGraphicsEffect(dropshadow)
-                self.setOpacity(InteractiveSquareGraphicsItem.MAX_OPACITY)
+                self.setOpacity(self.MAX_OPACITY)
             else:
                 self.setGraphicsEffect(None)
-                if self.opacity() != \
-                        InteractiveSquareGraphicsItem.HOVER_OPACITY:
-                    self.setOpacity(
-                        InteractiveSquareGraphicsItem.SELECTION_OPACITY)
+                if self.opacity() != self.HOVER_OPACITY:
+                    self.setOpacity(self.SELECTION_OPACITY)
         else:
             self.setGraphicsEffect(None)
-            self.setOpacity(InteractiveSquareGraphicsItem.MAX_OPACITY)
+            self.setOpacity(self.MAX_OPACITY)
 
     def paint(self, painter, option, widget=None):
         # Override the default selected appearance
@@ -715,720 +703,4 @@ class PythagorasTree:
         return Point(
             center.x + length / 2 * cos(angle),
             center.y + length / 2 * sin(angle)
-        )
-
-
-class TreeAdapter:
-    """Base class for tree representation.
-
-    Any subclass should implement the methods listed in this base class. Note
-    that some simple methods do not need to reimplemented e.g. is_leaf since
-    it that is the opposite of has_children.
-
-    """
-
-    ROOT_PARENT_INDEX = -1
-
-    def weight(self, node):
-        """Get the weight of the given node.
-
-        The weights of the children always sum up to 1.
-
-        Parameters
-        ----------
-        node : object
-            The label of the node.
-
-        Returns
-        -------
-        float
-            The weight of the node relative to its siblings.
-
-        """
-        raise NotImplemented()
-
-    def num_samples(self, node):
-        """Get the number of samples that a given node contains.
-
-        Parameters
-        ----------
-        node : object
-            A unique identifier of a node.
-
-        Returns
-        -------
-        int
-
-        """
-        raise NotImplemented()
-
-    def parent(self, node):
-        """Get the parent of a given node. Return -1 if the node is the root.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        object
-
-        """
-        raise NotImplemented()
-
-    def has_children(self, node):
-        """Check if the given node has any children.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        bool
-
-        """
-        raise NotImplemented()
-
-    def is_leaf(self, node):
-        """Check if the given node is a leaf node.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        object
-
-        """
-        return not self.has_children(node)
-
-    def children(self, node):
-        """Get all the children of a given node.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        Iterable[object]
-            A iterable object containing the labels of the child nodes.
-
-        """
-        raise NotImplemented()
-
-    def get_distribution(self, node):
-        """Get the distribution of types for a given node.
-
-        This may be the number of nodes that belong to each different classe in
-        a node.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        Iterable[int, ...]
-            The return type is an iterable with as many fields as there are
-            different classes in the given node. The values of the fields are
-            the number of nodes that belong to a given class inside the node.
-
-        """
-        raise NotImplemented()
-
-    def get_impurity(self, node):
-        """Get the impurity of a given node.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        object
-
-        """
-        raise NotImplemented()
-
-    def rules(self, node):
-        """Get a list of rules that define the given node.
-
-        Parameters
-        ----------
-        node : object
-
-        Returns
-        -------
-        Iterable[Rule]
-            A list of Rule objects, can be of any type.
-
-        """
-        raise NotImplemented()
-
-    def attribute(self, node):
-        """Get the attribute that splits the given tree.
-
-        Parameters
-        ----------
-        node
-
-        Returns
-        -------
-
-        """
-        raise NotImplemented()
-
-    def is_root(self, node):
-        """Check if a given node is the root node.
-
-        Parameters
-        ----------
-        node
-
-        Returns
-        -------
-
-        """
-        return node == self.root
-
-    def leaves(self, node):
-        """Get all the leavse that belong to the subtree of a given node.
-
-        Parameters
-        ----------
-        node
-
-        Returns
-        -------
-
-        """
-        raise NotImplemented()
-
-    def get_instances_in_nodes(self, dataset, nodes):
-        """Get all the instances belonging to a set of nodes for a given
-        dataset.
-
-        Parameters
-        ----------
-        dataset : Table
-            A Orange Table dataset.
-        nodes : iterable[TreeNode]
-            A list of tree nodes for which we want the instances.
-
-        Returns
-        -------
-
-        """
-        raise NotImplemented()
-
-    @property
-    def max_depth(self):
-        """Get the maximum depth that the tree reaches.
-
-        Returns
-        -------
-        int
-
-        """
-        raise NotImplemented()
-
-    @property
-    def num_nodes(self):
-        """Get the total number of nodes that the tree contains.
-
-        This does not mean the number of samples inside the entire tree, just
-        the number of nodes.
-
-        Returns
-        -------
-        int
-
-        """
-        raise NotImplemented()
-
-    @property
-    def root(self):
-        """Get the label of the root node.
-
-        Returns
-        -------
-        object
-
-        """
-        raise NotImplemented()
-
-    @property
-    def domain(self):
-        """Get the domain of the given tree.
-
-        The domain contains information about the classes what the tree
-        represents.
-
-        Returns
-        -------
-
-        """
-        raise NotImplemented()
-
-
-class SklTreeAdapter(TreeAdapter):
-    """SklTreeAdapter Class.
-
-    An abstraction on top of the scikit learn classification tree.
-
-    Parameters
-    ----------
-    tree : sklearn.tree
-        The raw sklearn classification tree.
-    domain : Orange.base.domain
-        The Orange domain that comes with the model.
-    adjust_weight : function, optional
-        If you want finer control over the weights of individual nodes you can
-        pass in a function that takes the existsing weight and modifies it.
-        The given function must have signture :: Number -> Number
-
-    """
-
-    def __init__(self, tree, domain, adjust_weight=lambda x: x):
-        self._tree = tree
-        self._domain = domain
-        self._adjust_weight = adjust_weight
-
-        # clear memoized functions
-        self.weight.cache_clear()
-        self._adjusted_child_weight.cache_clear()
-        self.parent.cache_clear()
-
-        self._all_leaves = None
-
-    @lru_cache()
-    def weight(self, node):
-        return self._adjust_weight(self.num_samples(node)) / \
-               self._adjusted_child_weight(self.parent(node))
-
-    @lru_cache()
-    def _adjusted_child_weight(self, node):
-        """Helps when dealing with adjusted weights.
-
-        It is needed when dealing with non linear weights e.g. when calculating
-        the log weight, the sum of logs of all the children will not be equal
-        to the log of all the data instances.
-        A simple example: log(2) + log(2) != log(4)
-
-        Parameters
-        ----------
-        node : int
-            The label of the node.
-
-        Returns
-        -------
-        float
-            The sum of all of the weights of the children of a given node.
-
-        """
-        return sum(self._adjust_weight(self.num_samples(c))
-                   for c in self.children(node)) \
-            if self.has_children(node) else 0
-
-    def num_samples(self, node):
-        return self._tree.n_node_samples[node]
-
-    @lru_cache()
-    def parent(self, node):
-        for children in (self._tree.children_left, self._tree.children_right):
-            try:
-                return (children == node).nonzero()[0][0]
-            except IndexError:
-                continue
-        return TreeAdapter.ROOT_PARENT_INDEX
-
-    def has_children(self, node):
-        return self._tree.children_left[node] != -1 \
-               or self._tree.children_right[node] != -1
-
-    def children(self, node):
-        if self.has_children(node):
-            return self._left_child(node), self._right_child(node)
-        return ()
-
-    def _left_child(self, node):
-        return self._tree.children_left[node]
-
-    def _right_child(self, node):
-        return self._tree.children_right[node]
-
-    def get_distribution(self, node):
-        return self._tree.value[node]
-
-    def get_impurity(self, node):
-        return self._tree.impurity[node]
-
-    @property
-    def max_depth(self):
-        return self._tree.max_depth
-
-    @property
-    def num_nodes(self):
-        return self._tree.node_count
-
-    @property
-    def root(self):
-        return 0
-
-    @property
-    def domain(self):
-        return self._domain
-
-    @lru_cache()
-    def rules(self, node):
-        if node != self.root:
-            parent = self.parent(node)
-            # Convert the parent list of rules into an ordered dict
-            pr = OrderedDict([(r.attr_name, r) for r in self.rules(parent)])
-
-            parent_attr = self.attribute(parent)
-            # Get the parent attribute type
-            parent_attr_cv = parent_attr.compute_value
-
-            is_left_child = self._left_child(parent) == node
-
-            # The parent split variable is discrete
-            if isinstance(parent_attr_cv, Indicator) and \
-                hasattr(parent_attr_cv.variable, 'values'):
-                values = parent_attr_cv.variable.values
-                attr_name = parent_attr_cv.variable.name
-                eq = not is_left_child * (len(values) != 2)
-                value = values[abs(parent_attr_cv.value -
-                                   is_left_child * (len(values) == 2))]
-                new_rule = DiscreteRule(attr_name, eq, value)
-                # Since discrete variables should appear in their own lines
-                # they must not be merged, so the dict key is set with the
-                # value, so the same keys can exist with different values
-                # e.g. #legs ≠ 2 and #legs ≠ 4
-                attr_name = attr_name + '_' + value
-            # The parent split variable is continuous
-            else:
-                attr_name = parent_attr.name
-                sign = not is_left_child
-                value = self._tree.threshold[self.parent(node)]
-                new_rule = ContinuousRule(attr_name, sign, value,
-                                          inclusive=is_left_child)
-
-            # Check if a rule with that attribute exists
-            if attr_name in pr:
-                pr[attr_name] = pr[attr_name].merge_with(new_rule)
-                pr.move_to_end(attr_name)
-            else:
-                pr[attr_name] = new_rule
-
-            return list(pr.values())
-        else:
-            return []
-
-    def attribute(self, node):
-        return self.domain.attributes[self.splitting_attribute(node)]
-
-    def splitting_attribute(self, node):
-        return self._tree.feature[node]
-
-    @lru_cache()
-    def leaves(self, node):
-        start, stop = self._subnode_range(node)
-        if start == stop:
-            # leaf
-            return np.array([node], dtype=int)
-        else:
-            is_leaf = self._tree.children_left[start:stop] == -1
-            assert np.flatnonzero(is_leaf).size > 0
-            return start + np.flatnonzero(is_leaf)
-
-    def _subnode_range(self, node):
-        """Get the range of indices where there are subnodes of the given node.
-        Taken from the classificationtreegraph.py"""
-
-        def find_largest_idx(n):
-            """It is necessary to locate the node with the largest index in the
-            children in order to get a good range. This is necessary with trees
-            that are not right aligned, which can happen when visualising
-            random forest trees."""
-            if self._tree.children_left[n] == -1:
-                return n
-
-            l_node = find_largest_idx(self._tree.children_left[n])
-            r_node = find_largest_idx(self._tree.children_right[n])
-
-            return l_node if l_node > r_node else r_node
-
-        right = left = node
-        if self._tree.children_left[left] == -1:
-            assert self._tree.children_right[node] == -1
-            return node, node
-        else:
-            left = self._tree.children_left[left]
-            right = find_largest_idx(right)
-
-            return left, right + 1
-
-    def get_samples_in_leaves(self, data):
-        """Get an array of instance indices that belong to each leaf.
-
-        For a given dataset X, separate the instances out into an array, so
-        they are grouped together based on what leaf they belong to.
-
-        Examples
-        --------
-        Given a tree with two leaf nodes ( A <- R -> B ) and the dataset X =
-        [ 10, 20, 30, 40, 50, 60 ], where 10, 20 and 40 belong to leaf A, and
-        the rest to leaf B, the following structure will be returned (where
-        array is the numpy array):
-        [array([ 0, 1, 3 ]), array([ 2, 4, 5 ])]
-
-        The first array represents the indices of the values that belong to the
-        first leaft, so calling X[ 0, 1, 3 ] = [ 10, 20, 40 ]
-
-        Parameters
-        ----------
-        data
-            A matrix containing the data instances.
-
-        Returns
-        -------
-        np.array
-            The indices of instances belonging to a given leaf.
-
-        """
-
-        def assign(node_id, indices):
-            if self._tree.children_left[node_id] == -1:
-                return [indices]
-            else:
-                feature_idx = self._tree.feature[node_id]
-                thresh = self._tree.threshold[node_id]
-
-                column = data[indices, feature_idx]
-                leftmask = column <= thresh
-                leftind = assign(self._tree.children_left[node_id],
-                                 indices[leftmask])
-                rightind = assign(self._tree.children_right[node_id],
-                                  indices[~leftmask])
-                return list.__iadd__(leftind, rightind)
-
-        # TODO this kind of cache can lead to all sorts of problems, but numpy
-        # arrays are unhashable, and this gives huge performance boosts
-        if self._all_leaves is not None:
-            return self._all_leaves
-
-        n, _ = data.shape
-
-        items = np.arange(n, dtype=int)
-        leaf_indices = assign(0, items)
-        self._all_leaves = leaf_indices
-        return leaf_indices
-
-    def get_instances_in_nodes(self, dataset, nodes):
-        if not isinstance(nodes, (list, tuple)):
-            nodes = [nodes]
-
-        node_leaves = [self.leaves(n.label) for n in nodes]
-        if len(node_leaves) > 0:
-            # get the leaves of the selected tree node
-            node_leaves = np.unique(np.hstack(node_leaves))
-
-            all_leaves = self.leaves(self.root)
-
-            indices = np.searchsorted(all_leaves, node_leaves, side='left')
-            # all the leaf samples for each leaf
-            leaf_samples = self.get_samples_in_leaves(dataset.X)
-            # filter out the leaf samples array that are not selected
-            leaf_samples = [leaf_samples[i] for i in indices]
-            indices = np.hstack(leaf_samples)
-        else:
-            indices = []
-
-        return dataset[indices] if len(indices) else None
-
-
-class Rule:
-    """The base Rule class for tree rules."""
-
-    def merge_with(self, rule):
-        """Merge the current rule with the given rule.
-
-        Parameters
-        ----------
-        rule : Rule
-
-        Returns
-        -------
-        Rule
-
-        """
-        raise NotImplemented()
-
-
-class DiscreteRule(Rule):
-    """Discrete rule class for handling Indicator rules.
-
-    Parameters
-    ----------
-    attr_name : str
-    eq : bool
-        Should indicate whether or not the rule equals the value or not.
-    value : object
-
-    Examples
-    --------
-    Age = 30
-    >>> rule = DiscreteRule('age', True, 30)
-    Name ≠ John
-    >>> rule = DiscreteRule('name', False, 'John')
-
-    Notes
-    -----
-      - Merging discrete rules is currently not implemented, the new rule is
-        simply returned and a warning is printed to stderr.
-
-    """
-
-    def __init__(self, attr_name, eq, value):
-        self.attr_name = attr_name
-        self.sign = eq
-        self.value = value
-
-    def merge_with(self, rule):
-        # It does not make sense to merge discrete rules, since they can only
-        # be eq or not eq.
-        from sys import stderr
-        print('WARNING: Merged two discrete rules `%s` and `%s`'
-              % (self, rule), file=stderr)
-        return rule
-
-    def __str__(self):
-        return '{} {} {}'.format(
-            self.attr_name, '=' if self.sign else '≠', self.value)
-
-
-class ContinuousRule(Rule):
-    """Continuous rule class for handling numeric rules.
-
-    Parameters
-    ----------
-    attr_name : str
-    gt : bool
-        Should indicate whether the variable must be greater than the value.
-    value : int
-    inclusive : bool, optional
-        Should the variable range include the value or not
-        (LT <> LTE | GT <> GTE). Default is False.
-
-    Examples
-    --------
-    x ≤ 30
-    >>> rule = ContinuousRule('age', False, 30, inclusive=True)
-    x > 30
-    >>> rule = ContinuousRule('age', True, 30)
-
-    Notes
-    -----
-      - Continuous rules can currently only be merged with other continuous
-        rules.
-
-    """
-
-    def __init__(self, attr_name, gt, value, inclusive=False):
-        self.attr_name = attr_name
-        self.sign = gt
-        self.value = value
-        self.inclusive = inclusive
-
-    def merge_with(self, rule):
-        if not isinstance(rule, ContinuousRule):
-            raise NotImplemented('Continuous rules can currently only be '
-                                 'merged with other continuous rules')
-        # Handle when both have same sign
-        if self.sign == rule.sign:
-            # When both are GT
-            if self.sign is True:
-                larger = self.value if self.value > rule.value else rule.value
-                return ContinuousRule(self.attr_name, self.sign, larger)
-            # When both are LT
-            else:
-                smaller = self.value if self.value < rule.value else rule.value
-                return ContinuousRule(self.attr_name, self.sign, smaller)
-        # When they have different signs we need to return an interval rule
-        else:
-            lt_rule = self if self.sign is False else rule
-            gt_rule = self if lt_rule != self else rule
-            return IntervalRule(self.attr_name, gt_rule, lt_rule)
-
-    def __str__(self):
-        return '%s %s %.3f' % (
-            self.attr_name, '>' if self.sign else '≤', self.value)
-
-
-class IntervalRule(Rule):
-    """Interval rule class for ranges of continuous values.
-
-    Parameters
-    ----------
-    attr_name : str
-    left_rule : ContinuousRule
-        The smaller (left) part of the interval.
-    right_rule : ContinuousRule
-        The larger (right) part of the interval.
-
-    Examples
-    --------
-    1 ≤ x < 3
-    >>> rule = IntervalRule('Rule',
-    >>>                     ContinuousRule('Rule', True, 1, inclusive=True),
-    >>>                     ContinuousRule('Rule', False, 3))
-
-    Notes
-    -----
-      - Currently, only cases which appear in classification and regression
-        trees are implemented. An interval can not be made up of two parts
-        (e.g. (-∞, -1) ∪ (1, ∞)).
-
-    """
-
-    def __init__(self, attr_name, left_rule, right_rule):
-        if not isinstance(left_rule, ContinuousRule):
-            raise AttributeError(
-                'The left rule must be an instance of the `ContinuousRule` '
-                'class.')
-        if not isinstance(right_rule, ContinuousRule):
-            raise AttributeError(
-                'The right rule must be an instance of the `ContinuousRule` '
-                'class.')
-
-        self.attr_name = attr_name
-        self.left_rule = left_rule
-        self.right_rule = right_rule
-
-    def merge_with(self, rule):
-        if isinstance(rule, ContinuousRule):
-            if rule.sign:
-                return IntervalRule(
-                    self.attr_name, self.left_rule.merge_with(rule),
-                    self.right_rule)
-            else:
-                return IntervalRule(
-                    self.attr_name, self.left_rule,
-                    self.right_rule.merge_with(rule))
-
-        elif isinstance(rule, IntervalRule):
-            return IntervalRule(
-                self.attr_name,
-                self.left_rule.merge_with(rule.left_rule),
-                self.right_rule.merge_with(rule.right_rule))
-
-    def __str__(self):
-        return '{} ∈ {}{:.3}, {:.3}{}'.format(
-            self.attr_name,
-            '[' if self.left_rule.inclusive else '(', self.left_rule.value,
-            self.right_rule.value, ']' if self.right_rule.inclusive else ')'
         )
